In release.sh, SRCBRANCH is set to something like base/head@rNNNNNN initially, and later prefixed with the value of SVNROOT in the form of:
	SRCBRANCH="${SVNROOT}${SRCBRANCH}"

Add trailing '/' to SVNROOT in the arm release configuration files
to follow suit, otherwise it is missing from the default svn path.

Ideally this would not be required, but this seems to be more sane
than forcibly suffixing SVNROOT with '/' or prefixing SRCBRANCH with
it, since _technically_, it should be perfectly valid for either to
be null.

Sponsored by:	The FreeBSD Foundation