Allow skipping the chroot buildworld/installworld/distribution.
The purpose of this is not to re-use the build environment, but
to allow skipping redundant stages if the build environment exists
via alternate means.

For snapshot builds, a single amd64 userland is built for all
non-i386 builds as part of a wrapper around release.sh, and the
resulting userland is installed into CHROOTDIR for each architecture.
Without this change, a buildworld would happen for each architecture
being built from the branch.

For head/, there are currently 12 individual builds across all
architectures, which means 12 repeated steps to set up the build
environment, all using the same svn revision, and all with the
same resulting userland.  Allowing a way to skip the redundant
provides a significant reduction in overall build time when building
more than one architecture release.  The wrapper script has been
patching release.sh to eliminate the redundancy, and this commit
eliminates the need to use any local patches to release.sh.

Sponsored by:	The FreeBSD Foundation